Updated plumbing for mdbook v0.41.1
mdbook v0.41.1 did a total overhaul of how the ToC is generated.
Instead of being statically included in every page, it's now
generated using javascript, AND written to its own file and
included as an iframe for browsers where javascript isn't being
executed. This means that the timestamp/commit info needs to be
added to three different files instead of just one.